+++ lib/ekylibre/tenant.rb
@@ -2,62 +2,52 @@ require 'apartment'
 require 'ekylibre/schema'
 
 module Ekylibre
-
   class TenantError < StandardError
   end
 
   class Tenant
-
-    AGGREGATION_NAME = "__all__"
+    AGGREGATION_NAME = '__all__'
 
     class << self
-
       # Tests existence of a tenant
       def exist?(name)
         list.include?(name)
       end
 
-
       # Tests existence of a tenant in DB
       # and removes it if not exist
       def check!(name, options = {})
         if list.include?(name)
-          unless Apartment.connection.schema_exists? name
-            drop(name, options)
-          end
+          drop(name, options) unless Apartment.connection.schema_exists? name
         end
       end
 
       # Returns the current tenant
       def current
         unless name = Apartment::Tenant.current
-          raise TenantError, "No current tenant"
+          fail TenantError, 'No current tenant'
         end
-        return name
+        name
       end
 
       # Returns the private directory of the current tenant
       # If tenant name given, it returns its private_directory
       def private_directory(name = nil)
-        Ekylibre.root.join("private", name || current)
+        Ekylibre.root.join('private', name || current)
       end
 
       # Create a new tenant with tables and co
       def create(name)
         name = name.to_s
         check!(name)
-        if exist?(name)
-          raise TenantError, "Already existing tenant"
-        end
+        fail TenantError, 'Already existing tenant' if exist?(name)
         Apartment::Tenant.create(name)
         add(name)
       end
 
       # Adds a tenant in config. No schema are created.
       def add(name)
-        unless list.include?(name)
-          list << name
-        end
+        list << name unless list.include?(name)
         write
       end
 
@@ -65,24 +55,16 @@ module Ekylibre
       # Nothing is done if already exist
       def setup!(name, options = {})
         check!(name, options)
-        unless exist?(name)
-          create(name)
-        end
+        create(name) unless exist?(name)
         switch!(name)
       end
 
       # Drop tenant
       def drop(name, options = {})
         name = name.to_s
-        unless exist?(name)
-          raise TenantError, "Unexistent tenant: #{name}"
-        end
-        if Apartment.connection.schema_exists? name
-          Apartment::Tenant.drop(name)
-        end
-        unless options[:keep_files]
-          FileUtils.rm_rf private_directory(name)
-        end
+        fail TenantError, "Unexistent tenant: #{name}" unless exist?(name)
+        Apartment::Tenant.drop(name) if Apartment.connection.schema_exists? name
+        FileUtils.rm_rf private_directory(name) unless options[:keep_files]
         @list[env].delete(name)
         write
       end
@@ -94,12 +76,9 @@ module Ekylibre
         end
       end
 
-
       def rename(old, new)
         check!(old)
-        unless exist?(old)
-          raise TenantError, "Unexistent tenant: #{name}"
-        end
+        fail TenantError, "Unexistent tenant: #{name}" unless exist?(old)
         ActiveRecord::Base.connection.execute("ALTER SCHEMA #{old.to_s.inspect} RENAME TO #{new.to_s.inspect};")
         @list[env].delete(old.to_s)
         @list[env] << new.to_s
@@ -107,7 +86,7 @@ module Ekylibre
 
       # Change current tenant
       def switch(name, &block)
-        raise "Need block to use Ekylibre::Tenant.switch" unless block_given?
+        fail 'Need block to use Ekylibre::Tenant.switch' unless block_given?
         Apartment::Tenant.switch(name, &block)
       end
 
@@ -115,18 +94,18 @@ module Ekylibre
         Apartment::Tenant.switch!(name)
       end
 
-      alias :current= :switch!
+      alias_method :current=, :switch!
 
       def switch_default!
         if list.empty?
-          raise TenantError, "No default tenant"
+          fail TenantError, 'No default tenant'
         else
           Apartment::Tenant.switch!(list.first)
         end
       end
 
       # Browse all tenant to make actions on it
-      def switch_each(&block)
+      def switch_each(&_block)
         list.each do |tenant|
           switch(tenant) do
             yield tenant
@@ -146,7 +125,7 @@ module Ekylibre
           @list ||= {}
         end
         @list[env] ||= []
-        return @list[env]
+        @list[env]
       end
 
       def drop_aggregation_schema!
@@ -158,9 +137,7 @@ module Ekylibre
       end
 
       def create_aggregation_views_schema!
-        if list.empty?
-          raise "No tenant to build an aggregation schema"
-        end
+        fail 'No tenant to build an aggregation schema' if list.empty?
         name = AGGREGATION_NAME
         connection = ActiveRecord::Base.connection
         connection.execute("CREATE SCHEMA IF NOT EXISTS #{name};")
@@ -168,21 +145,21 @@ module Ekylibre
           connection.execute "DROP VIEW IF EXISTS #{name}.#{table}"
           columns = Ekylibre::Schema.columns(table)
           queries = list.collect do |tenant|
-            "SELECT '#{tenant}' AS tenant_name, " + columns.collect{|c| c[:name] }.join(", ") + " FROM #{tenant}.#{table}"
+            "SELECT '#{tenant}' AS tenant_name, " + columns.collect { |c| c[:name] }.join(', ') + " FROM #{tenant}.#{table}"
           end
-          query = "CREATE VIEW #{name}.#{table} AS " + queries.join(" UNION ALL ")
+          query = "CREATE VIEW #{name}.#{table} AS " + queries.join(' UNION ALL ')
           connection.execute(query)
         end
       end
 
       def reset_search_path!
-        ActiveRecord::Base.connection.schema_search_path = Ekylibre::Application.config.database_configuration[::Rails.env]["schema_search_path"]
+        ActiveRecord::Base.connection.schema_search_path = Ekylibre::Application.config.database_configuration[::Rails.env]['schema_search_path']
       end
 
       private
 
       def config_file
-        Rails.root.join("config", "tenants.yml")
+        Rails.root.join('config', 'tenants.yml')
       end
 
       # Return the env
@@ -194,9 +171,6 @@ module Ekylibre
         FileUtils.mkdir_p(config_file.dirname)
         File.write(config_file, @list.to_yaml)
       end
-
     end
-
   end
-
 end
